+++ spec/api/endpoints/rounds_endpoint_spec.rb
@@ -3,7 +3,7 @@ require 'spec_helper'
 describe Api::Endpoints::RoundsEndpoint do
   include Api::Test::EndpointTest
 
-  let!(:team) { Fabricate(:team, api: true) }
+  let!(:team) { Fabricate(:team, api: true, sup_odd: false) }
 
   before do
     @cursor_params = { team_id: team.id.to_s }
@@ -13,12 +13,23 @@ describe Api::Endpoints::RoundsEndpoint do
   it_behaves_like 'a team token api', Round
 
   context 'round' do
-    let(:existing_round) { Fabricate(:round, team:) }
+    let(:last_round) { team.rounds.last }
+
+    before do
+      4.times { Fabricate(:user, team:) }
+      allow(team).to receive(:sync!)
+      allow_any_instance_of(Sup).to receive(:dm!)
+      team.sup!
+    end
 
     it 'returns a round' do
-      round = client.round(id: existing_round.id)
-      expect(round.id).to eq existing_round.id.to_s
-      expect(round._links.self._url).to eq "http://example.org/api/rounds/#{existing_round.id}"
+      round = client.round(id: last_round.id)
+      expect(round.id).to eq last_round.id.to_s
+      expect(round.paired_users_count).to eq 3
+      expect(round.paired_users.length).to eq 3
+      expect(round.missed_users_count).to eq 1
+      expect(round.missed_users.length).to eq 1
+      expect(round._links.self._url).to eq "http://example.org/api/rounds/#{last_round.id}"
     end
   end